chore: fix PR remarks
