+++ src/chart/components/pane/pane.component.ts
@@ -62,19 +62,19 @@ export class PaneComponent extends ChartBaseElement {
 
 	constructor(
 		public chartBaseModel: ChartBaseModel<'candle'>,
+		private mainCanvasModel: CanvasModel,
+		public readonly dynamicObjectsCanvasModel: CanvasModel,
 		private hitTestController: PaneHitTestController,
 		private config: FullChartConfig,
 		private mainScaleModel: ScaleModel,
 		private drawingManager: DrawingManager,
 		private chartPanComponent: ChartPanComponent,
-		private mainCanvasModel: CanvasModel,
 		private canvasInputListener: CanvasInputListenerComponent,
 		private canvasAnimation: CanvasAnimation,
 		private cursorHandler: CursorHandler,
 		public eventBus: EventBus,
 		private canvasBoundsContainer: CanvasBoundsContainer,
 		public readonly uuid: string,
-		public readonly dynamicObjectsCanvasModel: CanvasModel,
 		public seriesAddedSubject: Subject<DataSeriesModel>,
 		public seriesRemovedSubject: Subject<DataSeriesModel>,
 		options?: AtLeastOne<YExtentCreationOptions>,
